+++ flytekit/core/artifact.py
@@ -666,17 +666,20 @@ class Artifact(object):
             raise ValueError("Cannot create artifact id without name, project, domain, version")
         return self.to_flyte_idl().artifact_id
 
-    def embed_as_query(self, bindings: typing.List[Artifact]) -> art_id.ArtifactQuery:
+    def embed_as_query(
+        self, bindings: typing.List[Artifact], partition: Optional[str] = None, expr: Optional[str] = None
+    ) -> art_id.ArtifactQuery:
         """
         This should only be called in the context of a Trigger
         :param bindings: The list of artifacts in trigger_on
+        :param partition: Can embed a time partition
         :param expr: Only valid if there's a time partition.
         """
         # Find self in the list, raises ValueError if not there.
         idx = bindings.index(self)
         aq = art_id.ArtifactQuery(
             binding=art_id.ArtifactBindingData(
-                index=idx,
+                index=idx, partition_key=partition, transform=str(expr) if expr and partition else None
             )
         )
         return aq
